git-svn-id: http://svn.code.sf.net/p/gamescanner/code/trunk@147 f8d4d631-fb50-0410-a5e1-caf36a277498

+++ GameScanner/GameScanner/config.cpp
@@ -438,7 +438,7 @@ VOID WINAPI CFG_OnSelChanged(HWND hwndDlg)
 
 	UINT iSel = (UINT)tvitem.lParam;
 
-	if(g_currSelCfg==iSel)
+	if((UINT)g_currSelCfg==iSel)
 		return;
 
 	if(iSel==-1)
